+++ packages/contracts/contracts/openzeppelin-solidity/ReentrancyGuard.sol
@@ -11,15 +11,15 @@ contract ReentrancyGuard {
 
   /// @dev Constant for unlocked guard state - non-zero to prevent extra gas costs.
   /// See: https://github.com/OpenZeppelin/openzeppelin-solidity/issues/1056
-  uint private constant REENTRANCY_GUARD_FREE = 1;
+  uint256 private constant REENTRANCY_GUARD_FREE = 1;
 
   /// @dev Constant for locked guard state
-  uint private constant REENTRANCY_GUARD_LOCKED = 2;
+  uint256 private constant REENTRANCY_GUARD_LOCKED = 2;
 
   /**
    * @dev We use a single lock for the whole contract.
    */
-  uint private reentrancyLock = REENTRANCY_GUARD_FREE;
+  uint256 private reentrancyLock = REENTRANCY_GUARD_FREE;
 
   /**
    * @dev Prevents a contract from calling itself, directly or indirectly.
