+++ packages/contracts/contracts/storage/BZxObjects.sol
@@ -9,7 +9,7 @@ pragma solidity 0.5.2;
 contract BZxObjects {
 
     struct ListIndex {
-        uint index;
+        uint256 index;
         bool isSet;
     }
 
@@ -18,11 +18,11 @@ contract BZxObjects {
         address interestTokenAddress;
         address collateralTokenAddress;
         address oracleAddress;
-        uint loanTokenAmount;
-        uint interestAmount;
-        uint initialMarginAmount;
-        uint maintenanceMarginAmount;
-        uint maxDurationUnixTimestampSec;
+        uint256 loanTokenAmount;
+        uint256 interestAmount;
+        uint256 initialMarginAmount;
+        uint256 maintenanceMarginAmount;
+        uint256 maxDurationUnixTimestampSec;
         bytes32 loanOrderHash;
     }
 
@@ -31,10 +31,10 @@ contract BZxObjects {
         address takerAddress;
         address feeRecipientAddress;
         address tradeTokenToFillAddress;
-        uint lenderRelayFee;
-        uint traderRelayFee;
-        uint makerRole;
-        uint expirationUnixTimestampSec;
+        uint256 lenderRelayFee;
+        uint256 traderRelayFee;
+        uint256 makerRole;
+        uint256 expirationUnixTimestampSec;
         bool withdrawOnOpen;
         string description;
     }
@@ -43,27 +43,27 @@ contract BZxObjects {
         address trader;
         address collateralTokenAddressFilled;
         address positionTokenAddressFilled;
-        uint loanTokenAmountFilled;
-        uint loanTokenAmountUsed;
-        uint collateralTokenAmountFilled;
-        uint positionTokenAmountFilled;
-        uint loanStartUnixTimestampSec;
-        uint loanEndUnixTimestampSec;
+        uint256 loanTokenAmountFilled;
+        uint256 loanTokenAmountUsed;
+        uint256 collateralTokenAmountFilled;
+        uint256 positionTokenAmountFilled;
+        uint256 loanStartUnixTimestampSec;
+        uint256 loanEndUnixTimestampSec;
         bool active;
-        uint positionId;
+        uint256 positionId;
     }
 
     struct PositionRef {
         bytes32 loanOrderHash;
-        uint positionId;
+        uint256 positionId;
     }
 
     struct InterestData {
         address lender;
         address interestTokenAddress;
-        uint interestTotalAccrued;
-        uint interestPaidSoFar;
-        uint interestLastPaidDate;
+        uint256 interestTotalAccrued;
+        uint256 interestPaidSoFar;
+        uint256 interestLastPaidDate;
     }
 
 }
