+++ packages/contracts/contracts/tokens/EIP20Wrapper.sol
@@ -3,9 +3,9 @@ pragma solidity 0.5.2;
 
 
 interface NonCompliantEIP20 {
-    function transfer(address _to, uint _value) external;
-    function transferFrom(address _from, address _to, uint _value) external;
-    function approve(address _spender, uint _value) external;
+    function transfer(address _to, uint256 _value) external;
+    function transferFrom(address _from, address _to, uint256 _value) external;
+    function approve(address _spender, uint256 _value) external;
 }
